LPD-60291 Adapt liferay settings components to be reused
