+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/data-source/ConnectLiferayDXP.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import WizardPage, {Step} from 'settings/components/base-page/WizardPage';
-import {AssignLiferayIndividualsDataToChannelsStep} from 'settings/components/liferay/steps/AssignLiferayIndividualsDataToChannelsStep';
+import {Alert} from 'shared/types';
+import {AssignIndividualsDataToPropertiesStep} from 'settings/components/salesforce/steps/AssignIndividualsDataToChannelsStep';
 import {ConnectLiferayDXPStep} from 'settings/components/liferay/steps/ConnectLiferayDXPStep';
 import {ReviewSyncedDataStep} from 'settings/components/liferay/steps/ReviewSyncedDataStep';
+import {updateLiferay} from 'shared/api/data-source';
 
 const steps: Step[] = [
 	{
@@ -21,7 +23,30 @@ const steps: Step[] = [
 	},
 	{
 		content: props => (
-			<AssignLiferayIndividualsDataToChannelsStep {...props} />
+			<AssignIndividualsDataToPropertiesStep
+				{...props}
+				onSubmit={dataSource => {
+					if (
+						dataSource?.sitesSelected ||
+						dataSource?.contactsSelected
+					) {
+						props.addAlert({
+							alertType: Alert.Types.Success,
+							message: Liferay.Language.get(
+								'the-data-source-setup-is-now-complete,-and-you-will-begin-to-see-data-as-activities-occur-on-your-sites'
+							)
+						});
+					} else {
+						props.addAlert({
+							alertType: Alert.Types.Success,
+							message: Liferay.Language.get(
+								'the-data-source-setup-has-finished'
+							)
+						});
+					}
+				}}
+				updateDataSourceFn={updateLiferay}
+			/>
 		),
 		description: Liferay.Language.get(
 			'properties-allow-you-to-aggregate-data-on-your-users,-sites-and-dxp-commerce-channels.-individuals-data-will-be-available-in-any-property-they-are-assigned-to'
